fix(api): harden Neon pool and anchor streaks to the reader's day
node-pg returned BIGINT columns as strings and DATE (OID 1082) as JS Date, breaking read-time accumulation and the streak day keys. Coerce numerics, keep calendar-day strings via a type parser, and compute streak 'today' from a client-supplied ?day= so UTC servers do not flip a Tokyo reader's day. The pool now strips channel_binding (the pg driver cannot SCRAM-bind), retries once on transient connect errors after a 2s delay, and raises the connect timeout to absorb Neon cold starts.

+++ api/progress.js
@@ -69,7 +69,7 @@ module.exports = async (req, res) => {
 
           await client.query(
             `INSERT INTO progress (user_id, path, section_id, scroll_pct, read_seconds, completed, completed_at, answers, last_visited)
-             VALUES ($1, $2, $3, $4, $5, $6, CASE WHEN $6 THEN $8 ELSE NULL END, COALESCE($7::jsonb, '{}'::jsonb), $8)
+             VALUES ($1, $2, $3, $4, $5, $6, CASE WHEN $6 THEN $8::bigint ELSE NULL END, COALESCE($7::jsonb, '{}'::jsonb), $8::bigint)
              ON CONFLICT (user_id, path) DO UPDATE SET
                section_id = CASE WHEN EXCLUDED.section_id <> '' THEN EXCLUDED.section_id ELSE progress.section_id END,
                scroll_pct = GREATEST(progress.scroll_pct, EXCLUDED.scroll_pct),
